#32 Refactor utility classes into common
