+++ components/common/main/src/yi/component/common/ComparisonUtilities.java
@@ -1,4 +1,4 @@
-package yi.editor.utilities;
+package yi.component.common;
 
 /**
  * A collection of miscellaneous utilities.
